put resource/datasource routers in plugin_provider
