--- tfe/data_source_router.go
@@ -1,33 +1,0 @@
-package tfe
-
-import (
-	"context"
-	"log"
-
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-)
-
-type errUnsupportedDataSource string
-
-func (e errUnsupportedDataSource) Error() string {
-	return "unsupported data source: " + string(e)
-}
-
-type dataSourceRouter map[string]tfprotov5.DataSourceServer
-
-func (d dataSourceRouter) ValidateDataSourceConfig(ctx context.Context, req *tfprotov5.ValidateDataSourceConfigRequest) (*tfprotov5.ValidateDataSourceConfigResponse, error) {
-	ds, ok := d[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedDataSource(req.TypeName)
-	}
-	return ds.ValidateDataSourceConfig(ctx, req)
-}
-
-func (d dataSourceRouter) ReadDataSource(ctx context.Context, req *tfprotov5.ReadDataSourceRequest) (*tfprotov5.ReadDataSourceResponse, error) {
-	log.Printf("[DEBUG] @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@OMAR DATA SOURCE ROUTER READ")
-	ds, ok := d[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedDataSource(req.TypeName)
-	}
-	return ds.ReadDataSource(ctx, req)
-}
