--- tfe/resource_router.go
@@ -1,63 +1,0 @@
-package tfe
-
-import (
-	"context"
-
-	"github.com/hashicorp/terraform-plugin-go/tfprotov5"
-)
-
-type errUnsupportedResource string
-
-func (e errUnsupportedResource) Error() string {
-	return "unsupported resource: " + string(e)
-}
-
-type resourceRouter map[string]tfprotov5.ResourceServer
-
-func (r resourceRouter) ValidateResourceTypeConfig(ctx context.Context, req *tfprotov5.ValidateResourceTypeConfigRequest) (*tfprotov5.ValidateResourceTypeConfigResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.ValidateResourceTypeConfig(ctx, req)
-}
-
-func (r resourceRouter) UpgradeResourceState(ctx context.Context, req *tfprotov5.UpgradeResourceStateRequest) (*tfprotov5.UpgradeResourceStateResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.UpgradeResourceState(ctx, req)
-}
-
-func (r resourceRouter) ReadResource(ctx context.Context, req *tfprotov5.ReadResourceRequest) (*tfprotov5.ReadResourceResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.ReadResource(ctx, req)
-}
-
-func (r resourceRouter) PlanResourceChange(ctx context.Context, req *tfprotov5.PlanResourceChangeRequest) (*tfprotov5.PlanResourceChangeResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.PlanResourceChange(ctx, req)
-}
-
-func (r resourceRouter) ApplyResourceChange(ctx context.Context, req *tfprotov5.ApplyResourceChangeRequest) (*tfprotov5.ApplyResourceChangeResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.ApplyResourceChange(ctx, req)
-}
-
-func (r resourceRouter) ImportResourceState(ctx context.Context, req *tfprotov5.ImportResourceStateRequest) (*tfprotov5.ImportResourceStateResponse, error) {
-	res, ok := r[req.TypeName]
-	if !ok {
-		return nil, errUnsupportedResource(req.TypeName)
-	}
-	return res.ImportResourceState(ctx, req)
-}
